--- src/components/main/filterFrame.tsx
@@ -1,62 +1,0 @@
-import Image from "next/image";
-import RadioBox from "./radioBox";
-
-export default function FilterFrame() {
-  const cardData = [
-    {
-      id: 1,
-      path: "/images/infant.png",
-      text: "영유아",
-    },
-    {
-      id: 2,
-      path: "/images/child.png",
-      text: "어린이",
-    },
-    {
-      id: 3,
-      path: "/images/teen.png",
-      text: "청소년",
-    },
-  ];
-  return (
-    <section className="w-[screen] mt-[2rem]  flex items-center justify-center">
-      <div className="rounded-[0.625rem]">
-        <div className="flex flex-col items-center justify-center">
-          <text className="font-bold text-2xl ">
-            당신을 위한 맞춤형 필터링 서비스
-          </text>
-
-          <text className="pt-[1rem] text-xl font-medium">
-            우리 아이의 나이를 선택하고 필요한 정보를 받아보세요
-          </text>
-        </div>
-        <div className="flex gap-[3rem]">
-          <div className="bg-white w-[10.1rem] h-[15.5rem] mt-[3rem] drop-shadow-md rounded-[0.3rem] flex items-center justify-center flex-col hover:-translate-y-3 ease-in-out duration-300">
-            <text className="font-semibold font-xl">연령무관</text>
-          </div>
-          {cardData.map((item) => (
-            <div
-              key={item.id}
-              className="bg-white w-[10.1rem] h-[15.5rem] mt-[3rem] drop-shadow-md rounded-[0.3rem] flex items-center justify-center flex-col hover:-translate-y-3 ease-in-out duration-300"
-            >
-              <Image
-                className="h-[6.5rem] w-[5.2rem]"
-                src={item.path}
-                width={200}
-                height={200}
-                alt="로고"
-              />
-              <text className="font-semibold font-lg mt-[2rem]">
-                {item.text}
-              </text>
-            </div>
-          ))}
-          <div className="bg-yellowColor w-[27.5rem] h-[15.5rem] mt-[3rem]">
-            <RadioBox></RadioBox>
-          </div>
-        </div>
-      </div>
-    </section>
-  );
-}
